потом исправлю

+++ www/app/Http/Controllers/ReservationController.php
@@ -13,24 +13,24 @@ class ReservationController extends Controller
 {
     public function reserveBook(Request $request, $id): \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse|\Illuminate\Http\JsonResponse
     {
-        $date = now()->addDays(2)->format('Y-m-d');
+        if (Reserve::where('iduser', (Auth::user())->id)->count() == 0) {
+            $date = now()->addDays(2)->format('Y-m-d');
 
-        $book = Books::find($id);
-        if ($book->count > 0) {
-            $book->count = $book->count - 1;
-            $book->save();
+            $book = Books::find($id);
+            if ($book->count > 0) {
+                $book->count = $book->count - 1;
+                $book->save();
 
-            Reserve::create([
-                'idbook' => $id,
-                'iduser' => (Auth::user())->id,
-                'date' => $date,
-                'unique_key' => Auth::user()->unique_key
-            ]);
-
-            return redirect('/books');
-        } else {
-            return response()->json('false');
+                Reserve::create([
+                    'idbook' => $id,
+                    'iduser' => (Auth::user())->id,
+                    'date' => $date,
+                    'unique_key' => Auth::user()->unique_key
+                ]);
+            }
         }
+        return redirect('/books');
+
     }
 
     public function reserveBookForm(Request $request, $id): \Illuminate\Contracts\View\View
@@ -52,24 +52,26 @@ class ReservationController extends Controller
 
     public function issuanceReservedBook(Request $request): \Illuminate\Http\JsonResponse
     {
-        Reserve::destroy($request->get('reserve'));
+        if(Surrender::where('iduser', $request->get('user'))->count() < 0) {
+            Reserve::destroy($request->get('reserve'));
 
-        $date = now()->addDays(7)->format('Y-m-d');
-        Surrender::create([
-            'idbook' => $request->get('book'),
-            'iduser' => $request->get('user'),
-            'date' => $date,
-            'unique_key' => Auth::user()->unique_key
-        ]);
+            $date = now()->addDays(7)->format('Y-m-d');
+            Surrender::create([
+                'idbook' => $request->get('book'),
+                'iduser' => $request->get('user'),
+                'date' => $date,
+                'unique_key' => Auth::user()->unique_key
+            ]);
 
-        return response()->json('true');
+            return response()->json('true');
+        } else {
+            return response()->json('false');
+        }
     }
 
     public function deleteReservation(Request $request, $id): \Illuminate\Http\JsonResponse
     {
         Reserve::destroy($id);
-
-
 
         return response()->json('true');
     }
